requirements.txt.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
matplotlib==3.8.3
fpdf2==2.7.7